Add spokes-spinner coloring Rename `bubble-dotted` variant in TS types  to `spokes` Refactor `<OrbitProgress variant="spokes" />` for bug fix on inconsistent DOM manipulation
