+++ src/indicators/OrbitProgress/OrbitProgress.types.ts
@@ -4,9 +4,9 @@ export interface OrbitProgressProps extends CommonProps {
 	/**
 	 * variant of an orbitting loading indicator. Defaults to `disc`.
 	 */
-	variant?: "dotted" | "bubble-dotted" | "disc" | "split-disc" | "track-disc";
+	variant?: "dotted" | "spokes" | "disc" | "split-disc" | "track-disc";
 	/**
-	 * When used with variant `bubble-dotted` it sets a more densely packed animation.
+	 * When used with variant `spokes` it sets a more compact animation.
 	 */
 	dense?: boolean;
 }
