+++ src/indicators/OrbitProgress/Spokes/Spokes.types.ts
@@ -2,7 +2,7 @@ import { CommonProps } from "../../common.types";
 
 export interface SpokesProps extends CommonProps {
     /**
-	 * Make the animation more densely packed.
+	 * Make the animation more compact.
 	 */
 	dense?: boolean;
 }
